requirements.txt
discord.py-self
pyyaml
requests
pytz
aiohttp
pynacl